requirements.txt
requests
gradio
fastapi
uvicorn
TTS
python-dotenv
openai